# docs-site/openapi/relay.public.template.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Molii Public API",
    "version": "1.0.0",
    "description": "仅包含面向 API 用户的 Molii 创作接口。使用 Bearer API Key 鉴权；请勿将密钥置于浏览器或公开仓库。"
  },
  "servers": [{ "url": "http://127.0.0.1:3000", "description": "由 DOCS_API_BASE_URL 在生成时替换" }],
...
  "paths": {
...
    "/v1/assets": {
      "post": {
        "tags": ["Assets"], "summary": "创建临时素材资产", "operationId": "createAsset",
        "description": "通过公共 HTTP(S) `url` 创建可用于生成请求的临时资产。`url`、`asset_type`、`name` 必填；`asset_type` 为 `image`、`video` 或 `audio`，`name` 最长 80 个 Unicode 字符。成功响应只返回公开资产 `id`。资产有过期时间；上传或存储本身的费用以账户计费规则为准。429/5xx 可退避重试。",
        "security": [{ "BearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAssetRequest" }, "examples": { "success": { "value": { "url": "https://example.invalid/input.jpg", "asset_type": "image", "name": "reference-image" } } } } } },
        "responses": { "200": { "$ref": "#/components/responses/CreateAssetResult" }, "400": { "$ref": "#/components/responses/InvalidRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "429": { "$ref": "#/components/responses/RateLimited" }, "502": { "$ref": "#/components/responses/UpstreamFailure" } }
      }
    },
    "/v1/assets/{id}": {
      "get": {
        "tags": ["Assets"], "summary": "获取临时素材资产", "operationId": "getAsset",
        "description": "通过必填 `id` 获取当前 API Key 所属资产的安全元数据。已过期或不属于当前用户的资产不会返回。读取不增加生成计费；429/5xx 使用退避重试。",
        "security": [{ "BearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/AssetId" }],
        "responses": { "200": { "$ref": "#/components/responses/AssetResult" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" } }
      },
      "delete": {
        "tags": ["Assets"], "summary": "删除临时素材资产", "operationId": "deleteAsset",
        "description": "删除当前 API Key 所属的必填 `id` 资产。删除后不可恢复，不退还已发生的生成或存储费用；客户端可在网络超时时安全地再次确认资源是否存在，429/5xx 使用退避重试。",
        "security": [{ "BearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/AssetId" }],
        "responses": { "200": { "description": "删除成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteAssetResponse" }, "examples": { "success": { "value": { "success": true } } } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" } }
      }
    }
  },
  "components": {
    "securitySchemes": { "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "API key", "description": "Authorization: Bearer $MOLII_API_KEY" }, "AnthropicApiKey": { "type": "apiKey", "in": "header", "name": "x-api-key", "description": "Anthropic 兼容请求的 x-api-key 请求头。" }, "GeminiApiKey": { "type": "apiKey", "in": "header", "name": "x-goog-api-key", "description": "Gemini 兼容请求的 x-goog-api-key 请求头。" } },
    "parameters": {
...
      "AssetId": { "name": "id", "in": "path", "required": true, "description": "创建资产响应返回的资产 ID。", "schema": { "type": "string", "minLength": 1 }, "example": "asset_public_123" }
    },
    "responses": {
...
      "CreateAssetResult": { "description": "资产已创建", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAssetResponse" }, "examples": { "success": { "value": { "id": "asset_public_123" } } } } } },
      "AssetResult": { "description": "资产元数据", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Asset" }, "examples": { "success": { "value": { "id": "asset_public_123", "asset_type": "image", "name": "reference-image", "status": "ready", "created_at": 1720000000, "expires_at": 1720003600, "verified_at": 1720000001 } } } } } },
...
      "InvalidRequest": { "description": "请求参数无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "examples": { "invalid": { "value": { "error": { "code": "invalid_request", "message": "请检查必填字段和取值范围" } } } } } } },
      "Unauthorized": { "description": "缺少或无效的 API Key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "examples": { "unauthorized": { "value": { "error": { "code": "invalid_api_key", "message": "Authentication failed" } } } } } } },
      "NotFound": { "description": "资源不存在或无权访问", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "examples": { "notFound": { "value": { "error": { "code": "not_found", "message": "Resource not found" } } } } } } },
...
      "RateLimited": { "description": "请求频率受限；按 Retry-After 或指数退避后重试", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "examples": { "limited": { "value": { "error": { "code": "rate_limit_exceeded", "message": "Retry later" } } } } } } },
      "UpstreamFailure": { "description": "上游服务暂时不可用。对于可能计费的 POST，不要自动重放；使用公开 Request ID、日志和已知任务状态确认后，再决定后续操作。", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "examples": { "upstream": { "value": { "error": { "code": "upstream_error", "message": "Upstream service temporarily unavailable" } } } } } } }
    },
    "schemas": {
...
      "CreateAssetRequest": { "type": "object", "required": ["url", "asset_type", "name"], "properties": { "url": { "type": "string", "format": "uri", "pattern": "^https?://", "description": "公开 HTTP(S) URL" }, "asset_type": { "type": "string", "enum": ["image", "video", "audio"] }, "name": { "type": "string", "minLength": 1, "maxLength": 80 } } },
      "CreateAssetResponse": { "type": "object", "required": ["id"], "properties": { "id": { "type": "string" } } },
      "Asset": { "type": "object", "required": ["id", "asset_type", "status", "created_at", "expires_at", "verified_at"], "properties": { "id": { "type": "string" }, "asset_type": { "type": "string", "enum": ["image", "video", "audio"] }, "name": { "type": "string" }, "source_url": { "type": "string", "format": "uri" }, "preview_url": { "type": "string", "format": "uri" }, "source_kind": { "type": "string" }, "file_name": { "type": "string" }, "content_type": { "type": "string" }, "file_size": { "type": "integer" }, "status": { "type": "string" }, "created_at": { "type": "integer" }, "expires_at": { "type": "integer" }, "verified_at": { "type": "integer" } } },
      "DeleteAssetResponse": { "type": "object", "required": ["success"], "properties": { "success": { "type": "boolean", "const": true } } },
...
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "$ref": "#/components/schemas/ErrorDetail" } } },
      "ErrorDetail": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string" } } }
    }
  }
}